o4/app-dev-2/projects/o2_cinespoilers/requirements.txt
asgiref==3.11.1
Django==6.0.4
djangorestframework==3.17.1
sqlparse==0.5.5
